+++ tests/formatter/input-output-pairs/leading-and-trailing-wss-w-comments.out.gd
@@ -1,5 +1,5 @@
 class X:
 	func foo():
 		# aaa
-		pass  # bbb
-		# ccc
+		pass  #  bbb
+		# 	ccc